Src/backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==4.1.13
django-allauth==0.62.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
Markdown==3.6
PyJWT==2.8.0
pymongo==4.7.0
sqlparse==0.2.4
